+++ packages/reactabular-resizable/CHANGELOG.md
@@ -1,3 +1,9 @@
+8.3.0 / 2016-11-30
+==================
+
+  * Feature - Pass `index` to `helper` `getId`. The enhanced signature is `getId(column, index)`.
+  * Bug fix - Make sure `helper` merges class names correctly if there are multiple.
+
 8.0.0 / 2016-11-27
 ==================